demo/multimodal/offline/model_export/src/preprocessors/hf_extractor_preprocessor.requirements.txt
grpcio==1.45.0
grpcio-tools==1.45.0
pyarrow==6.0.1
requests==2.27.1
transformers==4.16.0
Pillow==9.1.0
